requirements.txt
numpy>=1.19.0
matplotlib>=3.3.0
pillow>=8.0.0